[SER-492] Move GetDefaults() to ConstellationBuildSettings class.
